src/apps/events/api/serializers.py
from rest_framework import serializers

from src.apps.events.models import Event, EventDate


class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = "__all__"


class EventDateSerializer(serializers.ModelSerializer):
    class Meta:
        model = EventDate
        fields = "__all__"
